requirements.txt
beautifulsoup4==4.11.1
discord.py==1.7.3
requests==2.26.0
transformers==4.18.0
